Refactoring invoices a little
Code was getting to crowded. Going to split folders into /home-data,pdf and /view will consist of the crud routes
