+++ client/src/components/invoices/deleteInvoice.js
@@ -9,7 +9,7 @@ const DeleteInvoice = (props) => {
     <div className='DeleteModal'>
       <div className='DeleteModal-Box'>
         <p> Are you sure you want to delete this? </p>
-        <Link to='/'>
+        <Link to='/invoices'>
           <button
             type='button'
             onClick={() => {
